Refactor backend to make finding code easier
Big refactor of the dataset router to match the same pattern used on the
frontend.  Methods realting to dataset are in `controllers/dataset.ts`,
those for revisions and fact tables are in `controllers/revision.ts`,
dimensions are in `controllers/dimension-processor.ts` and anything
relating to viewing or creating the cube is in
`controllers/cube-handler.ts`.
