+++ src/controllers/measure-handler.ts
@@ -2,6 +2,7 @@ import fs from 'fs';
 
 import { Database } from 'duckdb-async';
 import tmp from 'tmp';
+import { NextFunction, Request, Response } from 'express';
 
 import { LookupTable } from '../entities/dataset/lookup-table';
 import { FactTable } from '../entities/dataset/fact-table';
@@ -22,7 +23,12 @@ import { MeasureLookupTableExtractor } from '../extractors/measure-lookup-extrac
 import { DataValueFormat } from '../enums/data-value-format';
 import { ColumnDescriptor } from '../extractors/column-descriptor';
 import { LookupTableExtractor } from '../extractors/lookup-table-extractor';
+import { NotFoundException } from '../exceptions/not-found.exception';
+import { BadRequestException } from '../exceptions/bad-request.exception';
+import { getLatestRevision } from '../utils/latest';
+import { UnknownException } from '../exceptions/unknown.exception';
 
+import { uploadCSV } from './csv-processor';
 import { createFactTableQuery } from './cube-handler';
 
 async function cleanUpMeasure(measure: Measure) {
@@ -469,5 +475,100 @@ export const getMeasurePreview = async (dataset: Dataset, factTable: FactTable) 
         await quack.close();
         tempFile.removeCallback();
         throw error;
+    }
+};
+
+export const resetMeasure = async (req: Request, res: Response, next: NextFunction) => {
+    const dataset = res.locals.dataset;
+    const measure = dataset.measure;
+    if (!measure) {
+        next(new NotFoundException('errors.measure_missing'));
+        return;
+    }
+    logger.debug('Resetting measure by removing extractor, lookup table, info and join column');
+    measure.extractor = null;
+    if (measure.lookup) {
+        const measureLookupFilename = measure.lookup.filename;
+        const lookupTable: LookupTable = measure.lookupTable;
+        await lookupTable.remove();
+        measure.lookupTable = null;
+        logger.debug(`Removing file ${dataset.id}/${measureLookupFilename} from data lake`);
+        const datalakeService = new DataLakeService();
+        await datalakeService.deleteFile(measureLookupFilename, dataset.id);
+    }
+    if (measure.measureInfo) {
+        logger.debug('Removing all measure info');
+        for (const info of measure.measureInfo) {
+            await info.remove();
+        }
+    }
+    measure.joinColumn = null;
+    logger.debug('Saving measure and returning dataset');
+    await measure.save();
+    const updateDataset = await Dataset.findOneByOrFail({ id: dataset.id });
+    res.status(200);
+    const dto = DatasetDTO.fromDataset(updateDataset);
+    res.json(dto);
+};
+
+export const attachLookupTableToMeasure = async (req: Request, res: Response, next: NextFunction) => {
+    if (!req.file) {
+        next(new BadRequestException('errors.upload.no_csv'));
+        return;
+    }
+    const dataset: Dataset = res.locals.dataset;
+
+    // Replace calls that require this to calls that get a single factTable for all revisions to "present"
+    const factTable = getLatestRevision(dataset)?.factTables[0];
+    if (!factTable) {
+        next(new NotFoundException('errors.fact_table_invalid'));
+        return;
+    }
+    let fileImport: FactTable;
+    try {
+        fileImport = await uploadCSV(req.file.buffer, req.file?.mimetype, req.file?.originalname, res.locals.datasetId);
+    } catch (err) {
+        logger.error(`An error occurred trying to upload the file: ${err}`);
+        next(new UnknownException('errors.upload_error'));
+        return;
+    }
+
+    const tableMatcher = req.body as MeasureLookupPatchDTO;
+
+    try {
+        const result = await validateMeasureLookupTable(fileImport, factTable, dataset, req.file.buffer, tableMatcher);
+        if ((result as ViewErrDTO).status) {
+            const error = result as ViewErrDTO;
+            res.status(error.status);
+            res.json(result);
+            return;
+        }
+        res.status(200);
+        res.json(result);
+    } catch (err) {
+        logger.error(`An error occurred trying to handle measure lookup table with error: ${err}`);
+        next(new UnknownException('errors.upload_error'));
+    }
+};
+
+export const getPreviewOfMeasure = async (req: Request, res: Response, next: NextFunction) => {
+    const dataset = res.locals.dataset;
+    const factTable = getLatestRevision(dataset)?.factTables[0];
+    if (!dataset.measure) {
+        next(new NotFoundException('errors.measure_invalid'));
+        return;
+    }
+    if (!factTable) {
+        next(new NotFoundException('errors.fact_table_invalid'));
+        return;
+    }
+    try {
+        const preview = await getMeasurePreview(dataset, factTable);
+        res.status(200);
+        res.json(preview);
+    } catch (err) {
+        logger.error(`Something went wrong trying to get a preview of the dimension with the following error: ${err}`);
+        res.status(500);
+        res.json({ message: 'Something went wrong trying to generate a preview of the dimension' });
     }
 };
