+++ test/sample-files/csv/rowref-sw2-lookup.csv
@@ -1,0 +1,4 @@
+Description_en,Description_cy,Notes_en,Notes_cy,sort_order,RowRefAlt
+Health Visitor,Ymwelwyr iechyd,,,1,1
+Other Staff,Staff eraill,Contacts by other staff are only available from 2015-16 onwards.,au gan staff eraill ar gael.,3,3
+Wider Health Team,Y tîm iechyd ehangach,Not including health visitors,Ddim yn cynnwys ymwelwyr iechyd,2,2
